+++ app/src/main/java/com/kylecorry/trail_sense/tools/inventory/domain/PackItem.kt
@@ -3,8 +3,8 @@ package com.kylecorry.trail_sense.tools.inventory.domain
 import com.kylecorry.trailsensecore.domain.units.Weight
 
 data class PackItem(
-    val id: Int,
-    val packId: Int,
+    val id: Long,
+    val packId: Long,
     val name: String,
     val category: ItemCategory,
     val amount: Double = 0.0,
